[增加]1. 增加文件帮助类的注释和新接口

+++ Unity/Packages/com.jiangyin.unitygameframework@2021.5.31/Runtime/Helper/FileHelper.cs
@@ -6,19 +6,27 @@ using UnityEngine;
 
 namespace UnityGameFramework.Runtime
 {
+    /// <summary>
+    /// 文件帮助类
+    /// </summary>
     public static class FileHelper
     {
         /// <summary>
         /// 获取目录下的所有文件
         /// </summary>
-        /// <param name="files"></param>
-        /// <param name="dir"></param>
+        /// <param name="files">文件存放路径列表对象</param>
+        /// <param name="dir">目标目录</param>
         public static void GetAllFiles(List<string> files, string dir)
         {
-            string[] fls = Directory.GetFiles(dir);
-            foreach (string fl in fls)
+            if (!Directory.Exists(dir))
             {
-                files.Add(fl);
+                return;
+            }
+
+            string[] strings = Directory.GetFiles(dir);
+            foreach (string item in strings)
+            {
+                files.Add(item);
             }
 
             string[] subDirs = Directory.GetDirectories(dir);
@@ -31,9 +39,14 @@ namespace UnityGameFramework.Runtime
         /// <summary>
         /// 清理目录
         /// </summary>
-        /// <param name="dir"></param>
+        /// <param name="dir">目标路径</param>
         public static void CleanDirectory(string dir)
         {
+            if (!Directory.Exists(dir))
+            {
+                return;
+            }
+
             foreach (string subDir in Directory.GetDirectories(dir))
             {
                 Directory.Delete(subDir, true);
@@ -48,13 +61,13 @@ namespace UnityGameFramework.Runtime
         /// <summary>
         /// 目录复制
         /// </summary>
-        /// <param name="srcDir"></param>
-        /// <param name="tgtDir"></param>
+        /// <param name="srcDir">源路径</param>
+        /// <param name="targetDir">目标路径</param>
         /// <exception cref="Exception"></exception>
-        public static void CopyDirectory(string srcDir, string tgtDir)
+        public static void CopyDirectory(string srcDir, string targetDir)
         {
             DirectoryInfo source = new DirectoryInfo(srcDir);
-            DirectoryInfo target = new DirectoryInfo(tgtDir);
+            DirectoryInfo target = new DirectoryInfo(targetDir);
 
             if (target.FullName.StartsWith(source.FullName, StringComparison.CurrentCultureIgnoreCase))
             {
@@ -89,18 +102,23 @@ namespace UnityGameFramework.Runtime
         /// <summary>
         /// 复制文件到目标目录
         /// </summary>
-        /// <param name="sourceFileName"></param>
-        /// <param name="destFileName"></param>
+        /// <param name="sourceFileName">源路径</param>
+        /// <param name="destFileName">目标路径</param>
         /// <param name="overwrite">是否覆盖</param>
         public static void Copy(string sourceFileName, string destFileName, bool overwrite = false)
         {
+            if (!File.Exists(sourceFileName))
+            {
+                return;
+            }
+
             File.Copy(sourceFileName, destFileName, overwrite);
         }
 
         /// <summary>
         /// 删除文件
         /// </summary>
-        /// <param name="path"></param>
+        /// <param name="path">文件路径</param>
         public static void Delete(string path)
         {
             File.Delete(path);
@@ -109,7 +127,7 @@ namespace UnityGameFramework.Runtime
         /// <summary>
         /// 判断文件是否存在
         /// </summary>
-        /// <param name="path"></param>
+        /// <param name="path">文件路径</param>
         /// <returns></returns>
         public static bool IsExists(string path)
         {
@@ -139,10 +157,15 @@ namespace UnityGameFramework.Runtime
         /// <summary>
         /// 移动文件到目标目录
         /// </summary>
-        /// <param name="sourceFileName"></param>
-        /// <param name="destFileName"></param>
+        /// <param name="sourceFileName">文件源路径</param>
+        /// <param name="destFileName">目标路径</param>
         public static void Move(string sourceFileName, string destFileName)
         {
+            if (!File.Exists(sourceFileName))
+            {
+                return;
+            }
+
             Copy(sourceFileName, destFileName, true);
             Delete(sourceFileName);
         }
@@ -150,7 +173,7 @@ namespace UnityGameFramework.Runtime
         /// <summary>
         /// 读取指定路径的文件内容
         /// </summary>
-        /// <param name="path"></param>
+        /// <param name="path">文件路径</param>
         /// <returns></returns>
         public static byte[] ReadAllBytes(string path)
         {
@@ -165,28 +188,50 @@ namespace UnityGameFramework.Runtime
         /// <summary>
         /// 读取指定路径的文件内容
         /// </summary>
-        /// <param name="path"></param>
+        /// <param name="path">文件路径</param>
+        /// <param name="encoding">编码</param>
         /// <returns></returns>
-        public static string ReadAllText(string path)
+        public static string ReadAllText(string path, Encoding encoding)
         {
-            return File.ReadAllText(path);
+            return File.ReadAllText(path, encoding);
         }
 
         /// <summary>
         /// 读取指定路径的文件内容
         /// </summary>
-        /// <param name="path"></param>
+        /// <param name="path">文件路径</param>
+        /// <returns></returns>
+        public static string ReadAllText(string path)
+        {
+            return File.ReadAllText(path, Encoding.UTF8);
+        }
+
+        /// <summary>
+        /// 读取指定路径的文件内容
+        /// </summary>
+        /// <param name="path">文件路径</param>
+        /// <param name="encoding">编码</param>
+        /// <returns></returns>
+        public static string[] ReadAllLines(string path, Encoding encoding)
+        {
+            return File.ReadAllLines(path, encoding);
+        }
+
+        /// <summary>
+        /// 读取指定路径的文件内容
+        /// </summary>
+        /// <param name="path">文件路径</param>
         /// <returns></returns>
         public static string[] ReadAllLines(string path)
         {
-            return File.ReadAllLines(path);
+            return File.ReadAllLines(path, Encoding.UTF8);
         }
 
         /// <summary>
         /// 写入指定路径的文件内容
         /// </summary>
         /// <param name="path">文件路径</param>
-        /// <param name="buffer"></param>
+        /// <param name="buffer">写入内容</param>
         /// <returns></returns>
         public static void ReadAllLines(string path, byte[] buffer)
         {
@@ -197,18 +242,30 @@ namespace UnityGameFramework.Runtime
         /// 写入指定路径的文件内容
         /// </summary>
         /// <param name="path">文件路径</param>
-        /// <param name="lines"></param>
+        /// <param name="lines">写入的内容</param>
+        /// <param name="encoding">编码</param>
         /// <returns></returns>
-        public static void WriteAllLines(string path, string[] lines)
+        public static void WriteAllLines(string path, string[] lines, Encoding encoding)
         {
-            File.WriteAllLines(path, lines);
+            File.WriteAllLines(path, lines, encoding);
         }
 
         /// <summary>
         /// 写入指定路径的文件内容
         /// </summary>
         /// <param name="path">文件路径</param>
-        /// <param name="content"></param>
+        /// <param name="lines">写入的内容</param>
+        /// <returns></returns>
+        public static void WriteAllLines(string path, string[] lines)
+        {
+            File.WriteAllLines(path, lines, Encoding.UTF8);
+        }
+
+        /// <summary>
+        /// 写入指定路径的文件内容
+        /// </summary>
+        /// <param name="path">文件路径</param>
+        /// <param name="content">写入的内容</param>
         /// <param name="encoding">编码</param>
         /// <returns></returns>
         public static void WriteAllText(string path, string content, Encoding encoding)
@@ -217,14 +274,14 @@ namespace UnityGameFramework.Runtime
         }
 
         /// <summary>
-        /// 写入指定路径的文件内容
+        /// 写入指定路径的文件内容，UTF-8
         /// </summary>
         /// <param name="path">文件路径</param>
-        /// <param name="content"></param>
+        /// <param name="content">写入的内容</param>
         /// <returns></returns>
         public static void WriteAllText(string path, string content)
         {
-            File.WriteAllText(path, content);
+            File.WriteAllText(path, content, Encoding.UTF8);
         }
     }
 }
